Moved the error suppression context into its own folder
